Update class names & move styles.js -file up
